TallerApiRest_reto2/requirements.txt
Flask
pymongo
flask_swagger_ui
bcrypt
pyjwt
